src: M1 -pi atk/text; retype rock params to void* in pcompch.ch, text.ch
FossilOrigin-Name: 91d6da5d09087653990e6be745511fa7ba45e71e4fc7be192b28fa98b8b600e0

+++ src/atk/text/content.c
@@ -627,7 +627,7 @@ struct content *self;
     struct content_chapentry *cp;
     if(self->srctext == NULL) return;
     clear(self);
-    text_EnumerateEnvironments(self->srctext,0, text_GetLength(self->srctext),ns,(long)self);
+    text_EnumerateEnvironments(self->srctext,0, text_GetLength(self->srctext),ns,self);
     if(self->indexentry){
 	content_AlwaysInsertCharacters(self,content_GetLength(self) ,INDEXHEADER,strlen(INDEXHEADER));
     }
